+++ src/app/series/[name]/page.tsx
@@ -1,12 +1,10 @@
 import { type Metadata } from "next";
-import Link from "next/link";
 import { notFound } from "next/navigation";
 import ListHeading from "@/components/list-heading";
 import ListSection from "@/components/list-section";
-import TagList from "@/components/tag-list";
+import PostListOnSeries from "@/components/series/post-list-on-series";
 import { getPostsBySeries } from "@/service/posts";
 import { getSeries, getSeriesByName } from "@/service/series";
-import { formatDate } from "@/utils/date-formatter";
 
 type Props = {
   params: {
@@ -39,28 +37,7 @@ export default async function SeriesDetailPage({ params }: Props) {
           {series.description}
         </p>
       ) : null}
-      <ul className="flex flex-col gap-8">
-        {postsOnSeries.map((post, idx) => (
-          <li key={post.absoluteUrl}>
-            <div className="inline-block min-w-12 rounded-t-md border border-b-0 border-zinc-300 dark:border-zinc-700 px-2 py-0.5 font-bold text-2xl after:content-['.']">
-              {idx + 1}
-            </div>
-            <div className="p-2 flex flex-col gap-3 rounded-b-md border border-zinc-300 dark:border-zinc-700">
-              <Link
-                href={post.absoluteUrl}
-                className="flex flex-col gap-3 hover:text-indigo-500 transition-colors"
-              >
-                <h3 className="font-bold text-xl">{post.title}</h3>
-                {post.description.length > 0 ? <p>{post.description}</p> : null}
-              </Link>
-              <time className="text-sm text-zinc-700 dark:text-zinc-300">
-                {formatDate(post.createdAt)}
-              </time>
-              {post.tags ? <TagList tags={post.tags} /> : null}
-            </div>
-          </li>
-        ))}
-      </ul>
+      <PostListOnSeries postsOnSeries={postsOnSeries} />
     </ListSection>
   );
 }
